Move Elevator shuffle code to a new file
